[add] ワープの修正とBGM追加
