--- Assets/Prefabs/Fire.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 553b06a6616f55e46a86f3b695258411
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
