--- Assets/Scenes/FieldScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a004623ae7c2aef45ab56712902a8b8e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
