--- Assets/Scripts/InGame/Interact/WarpObject.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 4a78ffc34dd540f589dea969dcc7caaf
-timeCreated: 1749888753
