+++ Assets/StreamingAssets/BGM.acb.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 61c51bba925758e49b0643db722e8517
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
